apply afio-stdout patch from intranator repo
This was added to the Intranator repository in 2002 for reasons
that are unclear. Since the equivalent patch does not apply to
the 2.5.1 version of afio, rewrite it to convey the intent more
clearly.

[TJ: I remember vaguely this was due to speed reasons on our Celeron
processors back then as stderr is normally unbuffered and caused
a slow down when extracting a million files or so.]