add unit test for symlink guard
This adds a shell script to check the behavior of the symlink
handling modes during extraction.

Adapted and revised for the 2.5.2 release from an earlier version
of the directory traversal patches as developed at Intra2net.
It is unfortunate that upstream chose not to adopt our
differentiation between harmless and potentially harmful kinds of
symlinks (relative with ``..'', absolute) which forces us to drop
one entire category of tests.

Signed-off-by: Philipp Gesang <philipp.gesang@intra2net.com>